[UMA] moved unit tests to contrib
